Revert "Change datetime extraction functions to return INTEGER instead of BIGINT"
This reverts commit 4fb6eaa21daa3befdd97a41ac97142b3beccef9f.

Rationale: Canonical Trino behavior should remain BIGINT. PPL→Trino
compatibility must be solved at the SqlDialect layer, not by modifying
Trino's standard library. Per architectural review, OmniSqlDialect is
the correct place for MySQL/PPL→Trino semantic bridges.

Phase: omni-442 Phase 9.5 — architectural correction

+++ omni-engine/src/main/java/io/trino/operator/scalar/timestamp/ExtractDayOfWeek.java
@@ -30,14 +30,14 @@ public class ExtractDayOfWeek
     private ExtractDayOfWeek() {}
 
     @LiteralParameters("p")
-    @SqlType(StandardTypes.INTEGER)
+    @SqlType(StandardTypes.BIGINT)
     public static long extract(@SqlType("timestamp(p)") long timestamp)
     {
         return ISOChronology.getInstanceUTC().dayOfWeek().get(scaleEpochMicrosToMillis(timestamp));
     }
 
     @LiteralParameters("p")
-    @SqlType(StandardTypes.INTEGER)
+    @SqlType(StandardTypes.BIGINT)
     public static long extract(@SqlType("timestamp(p)") LongTimestamp timestamp)
     {
         return extract(timestamp.getEpochMicros());
